Streamlined the shape of the (temporary) mail object. Made an "envelope" which basically has signatures and addresses.
Wrote the sections that encrypt and decrypt mails.

print_mailbox should be displaying all the different folders in the 
user's mailbox.

process_mail recieves a raw mail from the server, decrypts and validates 
it before passing it onto a db function (that should be inserting it 
into the user's database)